% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ecotheory.R
\docType{package}
\name{ecotheory}
\alias{ecotheory}
\alias{ecotheory-package}
\title{ecotheory}
\description{
This is a compantion package to Vellend 201X. This package contains only
vignettes, which present detailed descriptions of code used in the book.
These vignettes reproduce most figures from the text.
}

